+++ src/common/pure/CowSwapSafeAppLink/index.tsx
@@ -1,0 +1,7 @@
+import { ExternalLink } from 'legacy/theme'
+
+import { SAFE_COW_APP_LINK } from 'common/constants/common'
+
+export function CowSwapSafeAppLink() {
+  return <ExternalLink href={SAFE_COW_APP_LINK}>CoW Swap Safe App↗</ExternalLink>
+}