+++ src/legacy/components/NumericalInput/index.tsx
@@ -46,13 +46,16 @@ const inputRegex = RegExp(`^\\d*(?:\\\\[.])?\\d*$`) // match escaped "." charact
 
 export const Input = React.memo(function InnerInput({
   value,
+  readOnly,
   onUserInput,
   placeholder,
   prependSymbol,
   type,
+  onFocus,
   ...rest
 }: {
   value: string | number
+  readOnly?: boolean
   onUserInput: (input: string) => void
   error?: boolean
   fontSize?: string
@@ -69,7 +72,11 @@ export const Input = React.memo(function InnerInput({
     <StyledInput
       {...rest}
       value={prependSymbol && value ? prependSymbol + value : value}
-      onFocus={autofocus}
+      readOnly={readOnly}
+      onFocus={(event) => {
+        autofocus(event)
+        onFocus?.(event)
+      }}
       onChange={(event) => {
         if (prependSymbol) {
           const value = event.target.value