+++ src/legacy/hooks/useRefetchPriceCallback.tsx
@@ -1,5 +1,7 @@
 import { useCallback } from 'react'
 
+import { PriceQuality } from '@cowprotocol/cow-sdk'
+
 import { useGetGpPriceStrategy } from 'legacy/hooks/useGetGpPriceStrategy'
 import { isOnline } from 'legacy/hooks/useIsOnline'
 import { AddGpUnsupportedTokenParams } from 'legacy/state/lists/actions'
@@ -226,8 +228,18 @@ export function useRefetchQuoteCallback() {
       }
 
       // Init get quote methods params
-      const bestQuoteParams = { ...params, strategy }
-      const fastQuoteParams = { quoteParams: { ...quoteParams, priceQuality: 'fast' } }
+
+      const bestQuoteParams = {
+        ...params,
+        strategy,
+        quoteParams,
+      }
+      const fastQuoteParams = {
+        quoteParams: {
+          ...quoteParams,
+          priceQuality: PriceQuality.FAST,
+        },
+      }
 
       // Register get best and fast quote methods on window
       registerOnWindow({