+++ src/legacy/state/price/updater.ts
@@ -16,11 +16,13 @@ import { useOrderValidTo } from 'legacy/state/user/hooks'
 import { isAddress } from 'legacy/utils'
 
 import { useIsEoaEthFlow } from 'modules/swap/hooks/useIsEoaEthFlow'
+import { useEnoughBalanceAndAllowance } from 'modules/tokens'
 import { useWalletInfo } from 'modules/wallet'
 
+import { getPriceQuality } from 'api/gnosisProtocol/api'
 import { LegacyFeeQuoteParams as LegacyFeeQuoteParamsFull } from 'api/gnosisProtocol/legacy/types'
+import { useVerifiedQuotesEnabled } from 'common/hooks/featureFlags/useVerifiedQuotesEnabled'
 import tryParseCurrencyAmount from 'lib/utils/tryParseCurrencyAmount'
-import { getAddress } from 'utils/getAddress'
 
 import { useAllQuotes, useIsBestQuoteLoading, useSetQuoteError } from './hooks'
 import { QuoteInformationObject } from './reducer'
@@ -119,14 +121,16 @@ function isRefetchQuoteRequired(
 
 export default function FeesUpdater(): null {
   const { chainId, account } = useWalletInfo()
+  const verifiedQuotesEnabled = useVerifiedQuotesEnabled(chainId)
 
   const { independentField, typedValue: rawTypedValue, recipient } = useSwapState()
   const {
     currencies: { INPUT: sellCurrency, OUTPUT: buyCurrency },
+    currenciesIds: { INPUT: sellCurrencyId, OUTPUT: buyCurrencyId },
+    parsedAmount,
   } = useDerivedSwapInfo()
 
-  const sellCurrencyId = getAddress(sellCurrency)?.toLowerCase()
-  const buyCurrencyId = getAddress(buyCurrency)?.toLowerCase()
+  const enoughBalance = useEnoughBalanceAndAllowance({ account, amount: parsedAmount })
 
   const { address: ensRecipientAddress } = useENSAddress(recipient)
   const receiver = ensRecipientAddress || recipient
@@ -203,6 +207,7 @@ export default function FeesUpdater(): null {
       userAddress: account,
       validTo,
       isEthFlow,
+      priceQuality: getPriceQuality({ verifyQuote: enoughBalance && verifiedQuotesEnabled }),
     }
 
     // Don't refetch if offline.
@@ -273,6 +278,8 @@ export default function FeesUpdater(): null {
     validTo,
     buyTokenAddressInvalid,
     sellTokenAddressInvalid,
+    enoughBalance,
+    verifiedQuotesEnabled,
   ])
 
   return null