+++ src/modules/tradeQuote/hooks/useQuoteParams.ts
@@ -1,14 +1,21 @@
 import { useMemo } from 'react'
 
+import { CurrencyAmount } from '@uniswap/sdk-core'
+
 import { NATIVE_CURRENCY_BUY_ADDRESS } from 'legacy/constants'
 
+import { useEnoughBalanceAndAllowance } from 'modules/tokens'
 import { useDerivedTradeState } from 'modules/trade/hooks/useDerivedTradeState'
 import { useWalletInfo } from 'modules/wallet'
 
+import { getPriceQuality } from 'api/gnosisProtocol/api'
+import { useVerifiedQuotesEnabled } from 'common/hooks/featureFlags/useVerifiedQuotesEnabled'
 import { getAddress } from 'utils/getAddress'
 
 export function useQuoteParams(amount: string | null) {
   const { chainId, account } = useWalletInfo()
+  const verifiedQuotesEnabled = useVerifiedQuotesEnabled(chainId)
+
   const { state } = useDerivedTradeState()
 
   const { inputCurrency, outputCurrency, orderKind } = state || {}
@@ -18,6 +25,11 @@ export function useQuoteParams(amount: string | null) {
   const fromDecimals = inputCurrency?.decimals
   const toDecimals = outputCurrency?.decimals
 
+  const enoughBalance = useEnoughBalanceAndAllowance({
+    account,
+    amount: (inputCurrency && amount && CurrencyAmount.fromRawAmount(inputCurrency, amount)) || undefined,
+  })
+
   return useMemo(() => {
     if (!sellToken || !buyToken || !amount) return
 
@@ -26,11 +38,25 @@ export function useQuoteParams(amount: string | null) {
       buyToken,
       amount,
       chainId,
+      // The field is used in _mapNewToLegacyParams()
+      userAddress: account,
       receiver: account,
       kind: orderKind,
       toDecimals,
       fromDecimals,
       isEthFlow: false,
+      priceQuality: getPriceQuality({ verifyQuote: enoughBalance && verifiedQuotesEnabled }),
     }
-  }, [sellToken, buyToken, toDecimals, fromDecimals, amount, account, chainId, orderKind])
+  }, [
+    amount,
+    account,
+    chainId,
+    orderKind,
+    enoughBalance,
+    buyToken,
+    fromDecimals,
+    sellToken,
+    toDecimals,
+    verifiedQuotesEnabled,
+  ])
 }