+++ src/modules/twap/containers/TwapFormWarnings/warnings/UnsupportedWalletWarning.tsx
@@ -1,7 +1,8 @@
 import { ExternalLink } from 'legacy/theme'
 
-import { SAFE_COW_APP_LINK, UNSUPPORTED_WALLET_LINK } from 'modules/twap/const'
+import { UNSUPPORTED_WALLET_LINK } from 'modules/twap/const'
 
+import { CowSwapSafeAppLink } from 'common/pure/CowSwapSafeAppLink'
 import { InlineBanner } from 'common/pure/InlineBanner'
 
 export function UnsupportedWalletWarning({ isSafeViaWc }: { isSafeViaWc: boolean }) {
@@ -11,7 +12,7 @@ export function UnsupportedWalletWarning({ isSafeViaWc }: { isSafeViaWc: boolean
         <strong>Use Safe web app</strong>
         <p>
           Use the Safe web app for advanced trading. <br />
-          Only available in the <ExternalLink href={SAFE_COW_APP_LINK}>CoW Swap Safe App↗</ExternalLink>
+          Only available in the <CowSwapSafeAppLink />
         </p>
       </InlineBanner>
     )